adk_melhor_proposta/pyproject.toml
[project]
name = "adk-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.1.1",
    "google-genai>=1.18.0",
    "python-dotenv>=1.1.0",
]
